Throw error if loadConfig cannot find file by path
If given a path to a file that doesn't exist, sass-lint should throw an
error to let the user know it cannot find the config file. Otherwise the
user will have no idea that the default config is being used and theirs
is being ignored.

When not given a config path, continue as usual: return an empty config.

This also fixes relative path issues in the failures test suite.