:bug: Fix a crash from empty-line-between-blocks
The tree traversal code that was checking for space between blocks in
`.sass` files was a little too lenient when checking where it should
traverse and was actually moving into the content of an `ident` node
when there wasn't an empty line after an `@include`. This looks like:

```sass
.item
  @include my-mixin
  color: blue
```

The tree traversal is looking for a `space` node and accidentally walks
into the `ident` node for `my-mixin`.

To prevent this, we now check to see if the parent is a node before
walking into it. If the parent isn't a node, then we should halt
execution and return a violation.

I have checked whether this problem happens in `.scss` files as well and
it appears it does not due to the presence of the braces for the blocks.
This leads me to believe that the problem is purely due to relying on
whitespace for delineate blocks.

Fixes #729
Fixes #793

`<DCO 1.1 Signed-off-by: Michael Herold michael@michaeljherold.com>`